backend/pyproject.toml
[project]
name = "english-agent"
version = "0.1.0"
description = "AI English Communication Agent - Voice & Text tutoring with LiveKit"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # LiveKit
    "livekit~=1.0",
    "livekit-agents[silero,turn-detector]~=1.0",

    # LangChain / LangGraph
    "langchain>=0.3.25",
    "langchain-core>=0.3.63",
    "langchain-google-genai>=2.1.5",
    "langchain-groq>=0.3.2",
    "langchain-openai>=0.3.19",
    "langgraph>=0.4.8",
    "langgraph-checkpoint-postgres>=2.0.21",

    # FastAPI
    "fastapi>=0.115.13",
    "uvicorn>=0.34.3",
    "python-multipart>=0.0.20",

    # Database
    "asyncpg>=0.30.0",
    "psycopg[binary,pool]>=3.2.9",
    "alembic>=1.16.2",
    "sqlalchemy[asyncio]>=2.0",

    # Memory
    "mem0ai>=0.1.115",

    # STT/TTS
    "faster-whisper>=1.1.0",
    "edge-tts>=7.0.0",

    # Config & Utils
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "loguru>=0.7.3",
    "httpx>=0.27",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
